Refactor Runner to separate out backend.
Prepare for adding new/multiple runners.
